requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
pillow
pyyaml